backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.1",
    "django-filter>=25.2",
    "django-structlog>=9.1.1",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "djoser>=2.3.3",
    "drf-spectacular>=0.28.0",
    "uuid>=1.30",
]

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
]
